Added workfiles for makesis 2 and 4 for unix.
git-svn-id: svn://localhost/public/MoSync/trunk@1459 6ad7a17e-d552-4a1e-ace4-f958fe14b117

+++ tools/makesis-2.0.0_unix/src/stub.cpp
@@ -53,7 +53,7 @@ HANDLE CreateFileA(LPCSTR filename, DWORD access, DWORD sharing,
 		fd = creat(filename, 0664);
 	}
 	else {
-		printf("CreateFileA: unknown creation flag %lu\n", creation);
+		printf("CreateFileA: unknown creation flag %lu\n", (long unsigned int)creation);
 	}
 
 	if(fd < 0) {
